Renamed some directories to prep for adding tests to git
