+++ .gitignore
@@ -1,4 +1,5 @@
 *_PRIVATE.py
+*_PRIV.*
 *_PRIVATE.tsqy
 *_PRIVATE.yaml
 *.bak
@@ -13,8 +14,8 @@ DC_Persistent/*
 VS_Figures/*
 VS_Persistent/*
 !VS_Persistent/figs_PUBL.pkl
-scripts/*
-!scripts/MAIN.py
+tests/*
+# !tests/MAIN.py
 
 Catalogs/Catalog*
 !Catalogs/Catalog_Audiobook.yaml
