+++ 00_🏠_Home.py
@@ -1,6 +1,6 @@
 import os
 import sys
 sys.path.append(os.curdir)
-from utils import Exhibit
+from scripts import Exhibit
 
 Exhibit.main()
