+++ pages/20_😴_Sleep.py
@@ -2,7 +2,6 @@
 This file is a Streamlit page.
 It is run via a call to `streamlit run` automatically due to its location in the /pages directory.
 """
-import datetime
 import streamlit as st
 # import pandas as pd
 # import matplotlib.pyplot as plt
@@ -12,8 +11,7 @@ import os
 import sys
 sys.path.append(os.curdir)
 import src.TimesheetGlobals as Global
-from utils import Visualize
-from utils import Exhibit
+from scripts import Visualize, Exhibit
 
 
 def main(path: str = os.path.join(Global.rootProjectPath(), 'VS_Persistent', 'figs_PUBL.pkl')):
